move and refactor store order tests
